added a timer creation feature, and checked that there was no conflict with what Christian came up with
