+++ README.md
@@ -1,77 +1,63 @@
-# Welcome!
+# Welcome to UHelper! (last updated on July 6 by Min)
 
-UHelper is an interactive bot that is designed to meet students' needs on Discord.
+UHelper is an interactive bot that is designed to meet students' needs on Discord. 
+Please use this Github link to learn more.
+https://github.com/christianjokhaji/UHelper
 
-Currently, we plan to implement the following features:
+## Overview
 
-Important Deadlines:
-July 11 - Phase 1
+The main goal of UHelper is to provide assistance with time management and wellness. 
+Users can create for tests or assignments at specified times and details. Once scheduled, 
+the bot will notify users individually in direct messages a day before the events. 
+For study assistance, they can input questions and answers for UHelper to quiz them later, 
+with optional hints available. To help boost students’ productivity and focus, 
+the bot can set up a Pomodoro timer, including customizable presets and collaborative features that 
+allow multiple users to have a timer in common to work remotely.
+
+The bot supports student wellness by suggesting healthy and handy cooking recipes based on given 
+ingredients. If requested, the bot can also display nutritional facts about each dish. Additionally,
+users can subscribe to receive daily motivational quotes and mental health tips. During breaks, 
+users can engage in mini-games, which track their records, to stay entertained. Finally, 
+to avoid command conflicts with other bots in the server, users can customize UHelper's command 
+lines. The customizability ensures a seamless and non-disruptive experience across various 
+functionalities.
+
+
+## Project Progresses
+I encourage you to use this section as a checklist for your work so that everyone can see how the
+project is going. You may add notes to communicate with your teammates. 
+
+Currently, we are working to implement the following features.
+
+- Timer
+  - [X] Create a timer preset by providing the amount of time wanted for work and break. (July 3)
+  - [ ] Initiate a timer instance by giving the bot the name of that timer.
+  - [ ] Notify users when time is up via DM.
+  - [ ] Cancel an ongoing timer.
+  - [ ] Invite other users to share a timer.
+
+- Minigames
+
+- Study help
+
+- Reminders for exams and assignments
+
+- Recipe
+
+- Motivational Quotes
+
+- Customizable Commands
+
+
+## Important Notes
+
+July 18 - Phase 1
+
 August 8 - Phase 2 & Final Presentation
 
 
-As a reminder, here are our user stories that we have to implement.
+## Blueprint
 
-Team Story: 
-I am a UofT student and want help keeping track of my schedule. I call the bot to create a schedule 
-and give it a set of important times to note down. The bot will then notify me in direct messages 
-at these times to help me manage my time. 
+https://docs.google.com/document/d/1OcYBGoSZbEqtA47CwSlzFe1wVuZo28Xl-FKUkS_0AUI/edit#heading=h.rwi1fv3j8vi2
 
-
-Mustafa’s Story: 
-I want to maintain my mental health and positively start my day. I call the bot to 
-receive daily motivational quotes and mental health tips.
-
-Example: 
-I want to receive a motivational quote every morning to help me start my day on a positive 
-note. I also want to get tips on managing stress and practicing mindfulness. I call the bot to 
-subscribe to daily quotes and tips. Every morning, the bot sends me a new motivational quote and 
-a mental health tip via direct message, helping me stay motivated and mentally healthy throughout 
-my studies.
-
-Aria’s Story: 
-I want to be given reminders of upcoming exams. I call the bot to create new exam events. 
-The chatbot sends me direct message notifications a day before my exam, showing me the exam's 
-subject, writing location, and time. 
-
-Dusan’s Story: 
-I want to call the bot for study help and give it a set of questions and answers for it to later 
-quiz me on. The bot can also provide me with hints, upon request.
-
-Example: 
-I am studying for the CSC207 midterm and want help remembering terms, so I write down questions, 
-answers, and respective hints and give them to the bot to later test me on. I then request the bot 
-using a command like !study, to state the questions back to me to test if I retained the knowledge, 
-giving hints if requested. The bot provides the various questions in a random order.
-
-Min's story: 
-I want to set up a Pomodoro study timer which will notify me at the end of each working session or 
-study break. The bot prompts me to select how many times I want the timer to repeat, and I can 
-enter a custom amount of time (in minutes) that I want to spend working, and also on break. 
-To promote collaboration within the server, I can invite others to share the same timer. 
-Users will be notified individually when the work/break is over via direct message. 
-At any point, I can cancel the timer.
-
-Xinyue(Cindy)’s story: 
-To stay healthy, I want to be able to call the bot for cooking help. The bot should provide recipes 
-and nutritional information for the food I have on hand, helping me quickly and easily make 
-nutritious meals.
-
-Example: 
-A user desires to know possible recipes for preparing a chicken and some vegetables, they then call 
-the bot to suggest related recipes. The bot can also show the user the nutrition facts about each 
-dish if requested.
-
-Christian’s story: 
-I want to play mini-games from a preselected list whenever I have a break/free time. I call the bot 
-to play a mini-game with me and we interact back-and-forth until the game is finished. It keeps 
-score against me and shows our record after each match. 
-
-Example: 
-!play {name_of_game}
-
-Story: 
-I have multiple bots in my server which use conflicting commands. I change the call command 
-for this bot to minimize interference. 
-
-Example: 
-Set the call as !c or !p 
+You'll find user stories and other details here.
