+++ src/main/java/ca/unknown/bot/app/Main.java
@@ -2,6 +2,7 @@ package ca.unknown.bot.app;
 
 import ca.unknown.bot.use_cases.EventListener;
 import ca.unknown.bot.use_cases.GameInteractor;
+import ca.unknown.bot.use_cases.TimerInteractor;
 import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.JDABuilder;
 import net.dv8tion.jda.api.interactions.commands.OptionType;
@@ -26,13 +27,23 @@ public class Main {
 
         // Adds use-case interactors to the bot instance.
         jda.addEventListener(new GameInteractor());
+        jda.addEventListener(new TimerInteractor());
+
 
         // Adds commands to the bot instance.
-        jda.updateCommands().addCommands(Commands.slash("rock-paper-scissors",
-                "Starts a game of rock paper scissors.")
+        jda.updateCommands().addCommands(
+                Commands.slash("rock-paper-scissors", "Starts a game of rock paper scissors.")
                 .addOptions(new OptionData(OptionType.STRING, "choice", "Rock, paper, or scissors.")
                         .addChoice("Rock", "rock")
                         .addChoice("Paper", "paper")
-                        .addChoice("Scissors", "scissors"))).queue();
+                        .addChoice("Scissors", "scissors")),
+                Commands.slash("preset", "Creates a new timer preset")
+                        .addOption(OptionType.INTEGER, "work", "how long a work session should be")
+                        .addOption(OptionType.INTEGER, "break", "how long a break should be")
+                        .addOption(OptionType.INTEGER, "iteration", "how many times you want a cycle to repeat")
+                        .addOption(OptionType.STRING, "name", "the name of the timer"),
+                Commands.slash("timer", "Initiates a Timer")
+                        .addOption(OptionType.STRING, "name", "the name of the timer instance")).queue();
+
     }
 }
