+++ src/main/java/ca/unknown/bot/entities/PomodoroFactory.java
@@ -1,4 +1,4 @@
-package ca.unknown.entities;
+package ca.unknown.bot.entities;
 
 public class PomodoroFactory implements PresetFactory {
     /**
