+++ src/main/java/ca/unknown/bot/entities/PresetFactory.java
@@ -1,4 +1,4 @@
-package ca.unknown.entities;
+package ca.unknown.bot.entities;
 
 public interface PresetFactory {
     Pomodoro create(Integer workMinute, Integer breakMinute, Integer iteration, String name);
