+++ src/main/java/ca/unknown/bot/miscellaneous/Test.java
@@ -1,0 +1,13 @@
+package ca.unknown.bot.miscellaneous;
+
+public class Test {
+    /**
+     * A test class for whatever you want to test.
+     *
+     *
+     * @param args
+     */
+    public static void main(String[] args) {
+
+            }
+    }
